feat/#276:서버 컴포넌트 변경
